feat: homepage SSR conversion and TypeScript fixes
- Convert app/page.tsx from client-side fetch to async server component
- Extract interactive UI to HomeClient.tsx (client component with match data as props)
- ISR revalidate=60 for homepage
- Fix TypeScript: ZodError.issues (was .errors, renamed in Zod v4)
- Fix TypeScript: Commentator model has no updatedAt field in sitemap.ts
- Fix TypeScript: explicit type annotation for filtered.map callback in dashboard/videos
